# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is API docs for Book API homework1",
    "version": "1.0.0",
    "title": "Book API"
  },
...
  "schemes": ["https", "http"],
  "paths": {
    "/book/:id": {
      "get": {
        "tags": ["book"],
        "summary": "send book id",
        "description": "get book by id",
        "operationId": "getBookById",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "bookId",
            "in": "path",
            "description": "ID of book to find",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": { "$ref": "#/definitions/ApiResponse" }
          }
        }
      },
      "delete": {
        "tags": ["book"],
        "summary": "delete book id",
        "operationId": "getBookById",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "bookId",
            "in": "path",
            "description": "ID of book to find",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": { "$ref": "#/definitions/ApiResponse" }
          }
        }
      }
    },
    "/book/list": {
      "get": {
        "tags": ["book"],
        "summary": "get all books",
        "description": "",
        "operationId": "listBooks",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [{}]
      }
    },
    "/book/create": {
      "get": {
        "tags": ["pet"],
        "summary": "create book",
        "description": "create book",
        "operationId": "addBook",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "body",
            "description": "id of book",
            "required": true,
            "type": "integer"
          },
          {
            "name": "title",
            "in": "body",
            "description": "title of book",
            "required": true,
            "type": "string"
          },
          {
            "name": "author",
            "in": "body",
            "description": "author of book",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": { "$ref": "#/definitions/Pet" }
            }
          },
          "400": { "description": "Invalid status value" }
        },
        "security": [{ "petstore_auth": ["write:pets", "read:pets"] }]
      }
    },
...
  },
  "securityDefinitions": {
    "api_key": { "type": "apiKey", "name": "api_key", "in": "header" },
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "https://petstore.swagger.io/oauth/authorize",
      "flow": "implicit",
      "scopes": {
        "read:pets": "read your pets",
        "write:pets": "modify pets in your account"
      }
    }
  },
  "definitions": {
    "ApiResponse": {
      "type": "object",
      "properties": {
        "code": { "type": "integer", "format": "int32" },
        "type": { "type": "string" },
        "message": { "type": "string" }
      }
    },
    "Category": {
      "type": "object",
      "properties": {
        "id": { "type": "integer", "format": "int64" },
        "name": { "type": "string" }
      },
      "xml": { "name": "Category" }
    },
    "Pet": {
      "type": "object",
      "required": ["name", "photoUrls"],
      "properties": {
        "id": { "type": "integer", "format": "int64" },
        "category": { "$ref": "#/definitions/Category" },
        "name": { "type": "string", "example": "doggie" },
        "photoUrls": {
          "type": "array",
          "xml": { "wrapped": true },
          "items": { "type": "string", "xml": { "name": "photoUrl" } }
        },
        "tags": {
          "type": "array",
          "xml": { "wrapped": true },
          "items": { "xml": { "name": "tag" }, "$ref": "#/definitions/Tag" }
        },
        "status": {
          "type": "string",
          "description": "pet status in the store",
          "enum": ["available", "pending", "sold"]
        }
      },
      "xml": { "name": "Pet" }
    },
    "Tag": {
      "type": "object",
      "properties": {
        "id": { "type": "integer", "format": "int64" },
        "name": { "type": "string" }
      },
      "xml": { "name": "Tag" }
    },
...
  },
...
}